Refactor Event into own package
