+++ server.go
@@ -1,6 +1,7 @@
 package tracker
 
 import (
+  "github.com/reevoo/tracker/event"
 	"github.com/reevoo/tracker/Godeps/_workspace/src/github.com/gin-gonic/gin"
 	"net/http"
 )
@@ -77,22 +78,22 @@ func (server Server) getStatus(context *gin.Context) {
 func (server Server) trackEvent(context *gin.Context) {
 	// Ensure the JSON is valid before returning
 	// context.BindJSON() sets an error status to the context on failure.
-	var event = NewEvent(
+	var e = event.New(
 		context.Request.URL.Query(),
 	)
 
-	if event.Empty() {
+	if e.Empty() {
 		context.String(http.StatusBadRequest, "No event params given.")
 	} else {
 		// We return the HTTP request quickly
 		// and process the event in the background.
-		go server.storeEvent(event)
+		go server.storeEvent(e)
 		context.String(http.StatusOK, "")
 	}
 }
 
-func (server Server) storeEvent(event Event) {
-	err := server.eventLogger.Log(event)
+func (server Server) storeEvent(e event.Event) {
+	err := server.eventLogger.Log(e)
 
 	if err != nil {
 		server.errorLogger.LogError(TrackerError{
